backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
Pillow>=10.0.0
trimesh>=4.0.0
numpy>=1.24.0
numpy-stl>=3.0.0
python-dotenv==1.0.0
gradio-client>=0.8.0
networkx>=3.0

